Revert "Fix gradle build caching (#2225)"
This reverts commit 23535798e8cc30e4349c5fe151953667ec72674f.
