+++ sqldelight-gradle-plugin/src/main/kotlin/com/squareup/sqldelight/gradle/SqlDelightTask.kt
@@ -31,7 +31,6 @@ import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.Internal
-import org.gradle.api.tasks.Nested
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.PathSensitive
 import org.gradle.api.tasks.PathSensitivity
@@ -56,7 +55,7 @@ abstract class SqlDelightTask : SqlDelightWorkerTask() {
   @Internal lateinit var sourceFolders: Iterable<File>
   @Internal lateinit var dependencySourceFolders: Iterable<File>
 
-  @Nested lateinit var properties: SqlDelightDatabasePropertiesImpl
+  @Input lateinit var properties: SqlDelightDatabaseProperties
 
   @Input var verifyMigrations: Boolean = false
