--- sqldelight-gradle-plugin/src/test/integration/settings.gradle
@@ -1,7 +1,0 @@
-rootProject.name = 'integration'
-
-buildCache {
-  local {
-    directory = new File(rootDir, 'build-cache')
-  }
-}
