+++ sqldelight-gradle-plugin/src/test/kotlin/com/squareup/sqldelight/tests/CompilationUnitTests.kt
@@ -2,9 +2,9 @@ package com.squareup.sqldelight.tests
 
 import com.alecstrong.sql.psi.core.DialectPreset
 import com.google.common.truth.Truth.assertThat
-import com.squareup.sqldelight.gradle.SqlDelightCompilationUnitImpl
-import com.squareup.sqldelight.gradle.SqlDelightDatabasePropertiesImpl
-import com.squareup.sqldelight.gradle.SqlDelightSourceFolderImpl
+import com.squareup.sqldelight.core.SqlDelightCompilationUnitImpl
+import com.squareup.sqldelight.core.SqlDelightDatabasePropertiesImpl
+import com.squareup.sqldelight.core.SqlDelightSourceFolderImpl
 import com.squareup.sqldelight.withTemporaryFixture
 import org.junit.Test
 import java.io.File
