+++ sqldelight-gradle-plugin/src/test/kotlin/com/squareup/sqldelight/tests/MultiModuleTests.kt
@@ -2,8 +2,8 @@ package com.squareup.sqldelight.tests
 
 import com.google.common.truth.Truth.assertThat
 import com.squareup.sqldelight.androidHome
-import com.squareup.sqldelight.gradle.SqlDelightCompilationUnitImpl
-import com.squareup.sqldelight.gradle.SqlDelightSourceFolderImpl
+import com.squareup.sqldelight.core.SqlDelightCompilationUnitImpl
+import com.squareup.sqldelight.core.SqlDelightSourceFolderImpl
 import com.squareup.sqldelight.properties
 import com.squareup.sqldelight.withInvariantPathSeparators
 import com.squareup.sqldelight.withSortedCompilationUnits
