+++ test-util/src/main/kotlin/com/squareup/sqldelight/test/util/TestEnvironment.kt
@@ -4,9 +4,7 @@ import com.alecstrong.sql.psi.core.DialectPreset
 import com.alecstrong.sql.psi.core.SqlAnnotationHolder
 import com.alecstrong.sql.psi.core.SqlCoreEnvironment
 import com.intellij.psi.PsiElement
-import com.squareup.sqldelight.core.SqlDelightCompilationUnit
-import com.squareup.sqldelight.core.SqlDelightDatabaseName
-import com.squareup.sqldelight.core.SqlDelightDatabaseProperties
+import com.squareup.sqldelight.core.SqlDelightDatabasePropertiesImpl
 import com.squareup.sqldelight.core.SqlDelightEnvironment
 import java.io.File
 
@@ -33,16 +31,16 @@ internal class TestEnvironment(
     val environment = SqlDelightEnvironment(
       sourceFolders = listOf(File(root)),
       dependencyFolders = emptyList(),
-      properties = object : SqlDelightDatabaseProperties {
-        override val packageName = "com.example"
-        override val className = "TestDatabase"
-        override val dependencies = emptyList<SqlDelightDatabaseName>()
-        override val compilationUnits = emptyList<SqlDelightCompilationUnit>()
-        override val outputDirectoryFile = outputDirectory
-        override val dialectPresetName = dialectPreset.name
-        override val deriveSchemaFromMigrations = this@TestEnvironment.deriveSchemaFromMigrations
-        override val rootDirectory = File(root)
-      },
+      properties = SqlDelightDatabasePropertiesImpl(
+        packageName = "com.example",
+        className = "TestDatabase",
+        dependencies = emptyList(),
+        compilationUnits = emptyList(),
+        outputDirectoryFile = outputDirectory,
+        dialectPresetName = dialectPreset.name,
+        deriveSchemaFromMigrations = deriveSchemaFromMigrations,
+        rootDirectory = File(root)
+      ),
       verifyMigrations = true,
       // hyphen in the name tests that our module name sanitizing works correctly
       moduleName = "test-module"
